Move caching fetched options to the abstract base class
