+++ src/sentry/static/sentry/app/components/group/externalIssueForm.tsx
@@ -8,7 +8,7 @@ import AbstractExternalIssueForm, {
 } from 'app/components/externalIssues/abstractExternalIssueForm';
 import NavTabs from 'app/components/navTabs';
 import {t, tct} from 'app/locale';
-import {Group, Integration, IntegrationExternalIssue, IssueConfigField} from 'app/types';
+import {Group, Integration, IntegrationExternalIssue} from 'app/types';
 import Form from 'app/views/settings/components/forms/form';
 
 const MESSAGES_BY_ACTION = {
@@ -127,9 +127,6 @@ export default class ExternalIssueForm extends AbstractExternalIssueForm<Props, 
   };
 
   renderBody() {
-    const {integrationDetails} = this.state;
-    const config: IssueConfigField[] =
-      (integrationDetails || {})[this.getConfigName()] || [];
-    return this.renderForm(config);
+    return this.renderForm(this.getCleanedFields());
   }
 }
